Separate out filesystem tests
